requirements.txt
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.23
numpy==1.19.5
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
